nvme_manager: decouple config dependency of PresentPin and PwrGoodPin
Nvme::read() does not allow power good checking if config just have
PwrGoodPin only available, for platform which doesn't have both
PresentPin and PwrGoodPin in config will not able to do power checking.

This patch move out GPIO pin reading to begin of each nvme read, and
set defaule value to PresentPin and PwrGoodPin if the pin is not
defined in the config file, so that the power checking can still be
performed without PresentPin defined.

default value of pins:
- PresentPin: IS_PRESENT (0)
- PwrGoodPin: POWERGD (1)

Signed-off-by: Potin Lai <potin.lai@quantatw.com>
Change-Id: I00fabf8f387c83bf139539bba9d5ba6aed693778